Add ParameterControlModule super class for control modules
